requirements.txt
accelerate==0.22.0
aiofiles==23.2.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.3.0
attrs==23.1.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
diffusers==0.20.2
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.103.1
ffmpy==0.3.1
filelock==3.9.0
fonttools==4.42.1
fsspec==2023.9.0
gradio==3.42.0
gradio_client==0.5.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.1
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.2.1
nest-asyncio==1.5.7
networkx==3.0
numpy==1.24.1
orjson==3.9.5
packaging==23.1
pandas==2.1.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.8.8
requests==2.28.1
rpds-py==0.10.0
safetensors==0.3.3
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
sympy==1.11.1
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.32.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.13
uvicorn==0.23.2
wcwidth==0.2.6
websockets==11.0.3
zipp==3.16.2
